refactor tests to be reusable across multiple target architectures
