+++ encoding/delta/binary_packed_amd64.go
@@ -3,7 +3,6 @@
 package delta
 
 import (
-	"github.com/segmentio/parquet-go/internal/bits"
 	"golang.org/x/sys/cpu"
 )
 
@@ -67,7 +66,7 @@ func encodeInt32AVX2(dst []byte, src []int32) []byte {
 	}
 
 	n := len(dst)
-	dst = resize(dst, n+maxHeaderLength)
+	dst = resize(dst, n+maxHeaderLength32)
 	dst = dst[:n+encodeBinaryPackedHeader(dst[n:], blockSize, numMiniBlocks, totalValues, int64(firstValue))]
 
 	if totalValues < 2 {
@@ -88,7 +87,7 @@ func encodeInt32AVX2(dst []byte, src []int32) []byte {
 		blockBitWidthsInt32AVX2(&bitWidths, &block)
 
 		n := len(dst)
-		dst = resize(dst, n+maxMiniBlockLength+16)
+		dst = resize(dst, n+maxMiniBlockLength32+16)
 		n += encodeBlockHeader(dst[n:], int64(minDelta), bitWidths)
 
 		for i, bitWidth := range bitWidths {
@@ -100,79 +99,6 @@ func encodeInt32AVX2(dst []byte, src []int32) []byte {
 		}
 
 		dst = dst[:n]
-	}
-
-	return dst
-}
-
-func (e *BinaryPackedEncoding) encodeInt64(dst []byte, src []int64) []byte {
-	totalValues := len(src)
-	firstValue := int64(0)
-	if totalValues > 0 {
-		firstValue = src[0]
-	}
-	dst = appendBinaryPackedHeader(dst, blockSize, numMiniBlocks, totalValues, firstValue)
-	if totalValues < 2 {
-		return dst
-	}
-
-	lastValue := firstValue
-	for i := 1; i < totalValues; {
-		block := make([]int64, blockSize)
-		block = block[:copy(block, src[i:])]
-		i += len(block)
-
-		for j, v := range block {
-			block[j], lastValue = v-lastValue, v
-		}
-
-		minDelta := bits.MinInt64(block)
-		bits.SubInt64(block, minDelta)
-
-		// blockSize x 8: we store at most `blockSize` count of values, which
-		// might be up to 64 bits in length, which is why we multiple by 8.
-		//
-		// Technically we could size the buffer to a smaller size when the
-		// bit width requires less than 8 bytes per value, but it would cause
-		// the buffer to be put on the heap since the compiler wouldn't know
-		// how much stack space it needs in advance.
-		miniBlock := make([]byte, blockSize*8)
-		bitWidths := [numMiniBlocks]byte{}
-		bitOffset := uint(0)
-		miniBlockLength := 0
-
-		for i := range bitWidths {
-			j := (i + 0) * miniBlockSize
-			k := (i + 1) * miniBlockSize
-
-			if k > len(block) {
-				k = len(block)
-			}
-
-			bitWidth := uint(bits.MaxLen64(block[j:k]))
-			if bitWidth != 0 {
-				bitWidths[i] = byte(bitWidth)
-
-				for _, bits := range block[j:k] {
-					for b := uint(0); b < bitWidth; b++ {
-						x := bitOffset / 8
-						y := bitOffset % 8
-						miniBlock[x] |= byte(((bits >> b) & 1) << y)
-						bitOffset++
-					}
-				}
-
-				miniBlockLength += (miniBlockSize * int(bitWidth)) / 8
-			}
-
-			if k == len(block) {
-				break
-			}
-		}
-
-		miniBlock = miniBlock[:miniBlockLength]
-		dst = appendBinaryPackedBlock(dst, int64(minDelta), bitWidths)
-		dst = append(dst, miniBlock...)
 	}
 
 	return dst
